+++ sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Blobs/src/StorageLoadBalancerQueue.cs
@@ -14,6 +14,7 @@ using Microsoft.Azure.WebJobs.Host.Executors;
 using Microsoft.Azure.WebJobs.Host.Listeners;
 using Microsoft.Azure.WebJobs.Host.Protocols;
 using Microsoft.Azure.WebJobs.Host.Queues;
+using Microsoft.Azure.WebJobs.Host.Scale;
 using Microsoft.Azure.WebJobs.Host.Timers;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
@@ -35,19 +36,22 @@ namespace Microsoft.Azure.WebJobs.Extensions.Storage.Blobs
         private readonly IWebJobsExceptionHandler _exceptionHandler;
         private readonly SharedQueueWatcher _sharedWatcher;
         private readonly QueueServiceClientProvider _queueServiceClientProvider;
+        private readonly ConcurrencyManager _concurrencyManager;
 
         public StorageLoadBalancerQueue(
                QueueServiceClientProvider queueServiceClientProvider,
                IOptions<QueuesOptions> queueOptions,
                IWebJobsExceptionHandler exceptionHandler,
                SharedQueueWatcher sharedWatcher,
-               ILoggerFactory loggerFactory)
+               ILoggerFactory loggerFactory,
+               ConcurrencyManager concurrencyManager)
         {
             _queueServiceClientProvider = queueServiceClientProvider;
             _queueOptions = queueOptions.Value;
             _exceptionHandler = exceptionHandler;
             _sharedWatcher = sharedWatcher;
             _loggerFactory = loggerFactory;
+            _concurrencyManager = concurrencyManager;
         }
 
         public IAsyncCollector<T> GetQueueWriter<T>(string queue)
@@ -118,7 +122,8 @@ namespace Microsoft.Azure.WebJobs.Extensions.Storage.Blobs
                 sharedWatcher: _sharedWatcher,
                 queueOptions: _queueOptions,
                 queueProcessor: queueProcessor,
-                functionDescriptor: new FunctionDescriptor { Id = SharedLoadBalancerQueueListenerFunctionId },
+                new FunctionDescriptor { Id = SharedLoadBalancerQueueListenerFunctionId },
+                _concurrencyManager,
                 maxPollingInterval: maxPollingInterval);
 
             return listener;